+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## master
+### Added
+- Feature: convert emoji names to Unicode emoji symbols
+
 ## [1.0.1] - 2018-01-20
 ### Added
 - Support for uppercase symbols